Blocks will be converted when the only inline styles prop is turned on.
